+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.2.0-beta6]
 - Fix event source check, prevent JS errors from externals messages
 - Support disposing apps
+- Support setting active app from client
 
 ## [2.2.0-beta5]
 - Fix bug that prevented app to start when it was too quick